+++ src/Modules/Expirator/Models/ExpirablePostModel.php
@@ -7,9 +7,9 @@
 namespace PublishPress\Future\Modules\Expirator\Models;
 
 use Closure;
+use PublishPress\Future\Framework\Logger\LoggerInterface;
 use PublishPress\Future\Framework\WordPress\Exceptions\NonexistentPostException;
 use PublishPress\Future\Framework\WordPress\Models\PostModel;
-use PublishPress\Future\Modules\Debug\DebugInterface;
 use PublishPress\Future\Modules\Expirator\ExpirationActionsAbstract;
 use PublishPress\Future\Modules\Expirator\HooksAbstract;
 use PublishPress\Future\Modules\Expirator\Interfaces\ExpirationActionInterface;
@@ -115,7 +115,6 @@ class ExpirablePostModel extends PostModel
 
     /**
      * @param int $postId
-     * @param \PublishPress\Future\Modules\Debug\DebugInterface $debug
      * @param \PublishPress\Future\Framework\WordPress\Facade\OptionsFacade $options
      * @param \PublishPress\Future\Framework\WordPress\Facade\HooksFacade $hooks
      * @param \PublishPress\Future\Framework\WordPress\Facade\UsersFacade $users
@@ -129,7 +128,6 @@ class ExpirablePostModel extends PostModel
      */
     public function __construct(
         $postId,
-        DebugInterface $debug,
         $options,
         $hooks,
         $users,
@@ -139,12 +137,13 @@ class ExpirablePostModel extends PostModel
         $termModelFactory,
         $expirationActionFactory,
         $actionArgsModelFactory,
-        $defaultDataModelFactory
+        $defaultDataModelFactory,
+        LoggerInterface $logger
     ) {
-        parent::__construct($postId, $termModelFactory, $debug, $hooks);
+        parent::__construct($postId, $termModelFactory, $hooks, $logger);
 
         $this->postId = $postId;
-        $this->debug = $debug;
+        $this->logger = $logger;
         $this->options = $options;
         $this->scheduler = $scheduler;
         $this->users = $users;
@@ -413,13 +412,13 @@ class ExpirablePostModel extends PostModel
         $postId = $this->getPostId();
 
         if (! $this->isExpirationEnabled() && ! $force) {
-            $this->debug->log($postId . ' -> Tried to run action but future action is NOT ACTIVATED for the post');
+            $this->logger->debug($postId . ' -> Tried to run action but future action is NOT ACTIVATED for the post');
 
             return false;
         }
 
         if (! $this->isExpirationEnabled() && $force) {
-            $this->debug->log(
+            $this->logger->debug(
                 $postId . ' -> Future action is not activated for the post, but $force = true'
             );
         }
@@ -434,13 +433,13 @@ class ExpirablePostModel extends PostModel
         $expirationAction = $this->getExpirationAction();
 
         if (! $expirationAction) {
-            $this->debug->log($postId . ' -> Future action cancelled, expiration action is not found');
+            $this->logger->debug($postId . ' -> Future action cancelled, expiration action is not found');
 
             return false;
         }
 
         if (! $expirationAction instanceof ExpirationActionInterface) {
-            $this->debug->log($postId . ' -> Future action cancelled, expiration action is not valid');
+            $this->logger->debug($postId . ' -> Future action cancelled, expiration action is not valid');
 
             return false;
         }
@@ -448,13 +447,13 @@ class ExpirablePostModel extends PostModel
         $result = $expirationAction->execute();
 
         if (! is_bool($result)) {
-            $this->debug->log($postId . ' -> ACTION ' . $expirationAction . ' returned a non boolean value');
+            $this->logger->debug($postId . ' -> ACTION ' . $expirationAction . ' returned a non boolean value');
 
             return false;
         }
 
         if (! $result) {
-            $this->debug->log(
+            $this->logger->debug(
                 // phpcs:ignore WordPress.PHP.DevelopmentFunctions.error_log_print_r
                 $postId . ' -> FAILED ' . print_r($this->getExpirationDataAsArray(), true)
             );
@@ -462,7 +461,7 @@ class ExpirablePostModel extends PostModel
             return false;
         }
 
-        $this->debug->log(
+        $this->logger->debug(
             // phpcs:ignore WordPress.PHP.DevelopmentFunctions.error_log_print_r
             $postId . ' -> PROCESSED ' . print_r($this->getExpirationDataAsArray(), true)
         );
@@ -636,7 +635,7 @@ class ExpirablePostModel extends PostModel
         );
 
         if (empty($emailBody)) {
-            $this->debug->log($this->getPostId() . ' -> Tried to send email, but notification text is empty');
+            $this->logger->debug($this->getPostId() . ' -> Tried to send email, but notification text is empty');
 
             return false;
         }
@@ -773,12 +772,12 @@ class ExpirablePostModel extends PostModel
         $emailSent = false;
 
         if (! empty($emailAddresses)) {
-            $this->debug->log($this->getPostId() . ' -> SENDING EMAIL TO (' . implode(', ', $emailAddresses) . ')');
+            $this->logger->debug($this->getPostId() . ' -> SENDING EMAIL TO (' . implode(', ', $emailAddresses) . ')');
 
             // Send each email.
             foreach ($emailAddresses as $email) {
                 if (empty($email)) {
-                    $this->debug->log($this->getPostId() . ' -> EMPTY EMAIL ADDRESS, SKIPPING');
+                    $this->logger->debug($this->getPostId() . ' -> EMPTY EMAIL ADDRESS, SKIPPING');
 
                     continue;
                 }
@@ -791,7 +790,7 @@ class ExpirablePostModel extends PostModel
                     $emailAttachments
                 );
 
-                $this->debug->log(
+                $this->logger->debug(
                     sprintf(
                         '%d -> %s (%s)',
                         $this->getPostId(),